+++ cryptography/hazmat/primitives/asymmetric/padding.py
@@ -20,3 +20,8 @@ from cryptography.hazmat.primitives import interfaces
 @utils.register_interface(interfaces.AsymmetricPadding)
 class PKCS1v15(object):
     name = "EMSA-PKCS1-v1_5"
+
+
+@utils.register_interface(interfaces.AsymmetricPadding)
+class PSS(object):
+    name = "EMSA-PSS"